Add go 1.7 build tags for all the Run code
Since go 1.6 did not have a Run method on *testing.T, we need to put build tags on that code.
