+++ run_test.go
@@ -1,3 +1,5 @@
+// +build go1.7
+
 package expect_test
 
 import (
